brew: better check if cask is installed
Cask updates will cause `brew cask list xxx` to report the cask is not
installed, even though it is, but on a version prior to the current.

So instead, check if the cask has a `/opt/homebrew-cask/Caskroom/xxx`
directory and if so log a warning if `brew cask list xxx` reports error.